Separate fortnite shop to two images: track and non-tracks

+++ src/main/java/bobo/utils/DailyTask.java
@@ -1,10 +1,10 @@
 package bobo.utils;
 
 import bobo.Bobo;
-import net.dv8tion.jda.api.entities.channel.middleman.GuildChannel;
 import net.dv8tion.jda.api.entities.channel.middleman.GuildMessageChannel;
 import net.dv8tion.jda.api.utils.FileUpload;
 
+import java.awt.image.BufferedImage;
 import java.io.File;
 import java.sql.*;
 import java.time.Duration;
@@ -12,9 +12,12 @@ import java.time.LocalDateTime;
 import java.time.ZoneId;
 import java.time.ZonedDateTime;
 import java.time.format.DateTimeFormatter;
+import java.util.List;
+import java.util.Objects;
 import java.util.concurrent.Executors;
 import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.TimeUnit;
+import java.util.stream.Collectors;
 
 import static bobo.commands.general.FortniteCommand.convertBufferedImageToFile;
 
@@ -49,32 +52,42 @@ public class DailyTask {
      * Sends the Fortnite shop image to all registered Discord channels.
      */
     public void sendFortniteShopImage() {
-        GuildChannel channel;
         ZonedDateTime nowInUTC = ZonedDateTime.now(ZoneId.of("UTC"));
         DateTimeFormatter formatter = DateTimeFormatter.ofPattern("EEEE, MMMM d, yyyy");
-        final String message = "# Fortnite Item Shop" + "\n" + "## " + nowInUTC.format(formatter);
+        final String message = "# Fortnite Shop" + "\n" + "## " + nowInUTC.format(formatter);
 
         try (Connection connection = SQLConnection.getConnection();
-             Statement statement = connection.createStatement()) {
-            ResultSet resultSet = statement.executeQuery(selectFortniteChannelsSQL);
+             Statement statement = connection.createStatement();
+             ResultSet resultSet = statement.executeQuery(selectFortniteChannelsSQL)) {
 
-            File file = convertBufferedImageToFile(FortniteAPI.getShopImage(), "shop");
-            FileUpload upload;
-            if (file != null) {
-                upload = FileUpload.fromData(file);
-            } else {
+            List<BufferedImage> images = FortniteAPI.getShopImages();
+            if (images == null) {
                 return;
             }
 
+            List<File> files = images.stream()
+                    .map(image -> convertBufferedImageToFile(image, "shop"))
+                    .filter(Objects::nonNull)
+                    .toList();
+
+            if (files.isEmpty()) {
+                return;
+            }
+
+            List<FileUpload> fileUploads = files.stream()
+                    .map(FileUpload::fromData)
+                    .collect(Collectors.toList());
+
             while (resultSet.next()) {
-                channel = Bobo.getJDA().getGuildChannelById(resultSet.getString("channel_id"));
-                GuildMessageChannel messageChannel = (GuildMessageChannel) channel;
+                GuildMessageChannel messageChannel = (GuildMessageChannel) Bobo.getJDA().getGuildChannelById(resultSet.getString("channel_id"));
                 if (messageChannel != null) {
-                    messageChannel.sendMessage(message).addFiles(upload).queue(success -> {
-                        if (!file.delete()) {
-                            System.err.println("Failed to delete file: " + file.getAbsolutePath());
-                        }
-                    });
+                    messageChannel.sendMessage(message)
+                            .setFiles(fileUploads)
+                            .queue(success -> files.forEach(file -> {
+                                if (!file.delete()) {
+                                    System.err.println("Failed to delete file: " + file.getAbsolutePath());
+                                }
+                            }));
                 }
             }
         } catch (SQLException e) {
